Add files for SQL query in SSMS
The file "Query for display data.sql" contains SQL queries intended to be executed in SQL Server Management Studio (SSMS) to retrieve and display data based on specified criteria. This file serves as a set of commands to extract specific information from the database. On the other hand, the file "Table Relational.png" illustrates the relational structure of tables within the database. It provides a visual representation of how different tables are interlinked, showcasing the relationships between them. Lastly, the "create+insert.bak" file is crucial for restoring the database in SSMS. It contains the necessary SQL commands for both creating tables and inserting initial data. Importing or restoring this ".bak" file into SSMS ensures that the data is available for further processing and analysis within the SQL Server environment.